@@ -1,24 +1,28 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 31
     defaultConfig {
         //https://stackoverflow.com/questions/46267621/unable-to-merge-dex
         multiDexEnabled true
 
         applicationId "com.example.prabin.agriculturearcgis"
-        minSdkVersion 21
-        targetSdkVersion 26
+        minSdk 23
+        targetSdkVersion 31
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
 
+//    packagingOptions {
+//        exclude 'META-INF/LGPL2.1'
+//        exclude 'META-INF/LICENSE'
+//        exclude 'META-INF/NOTICE'
+//    }
     packagingOptions {
-        exclude 'META-INF/LGPL2.1'
-        exclude 'META-INF/LICENSE'
-        exclude 'META-INF/NOTICE'
+        resources.excludes.add("META-INF/*")
     }
+
     buildTypes {
         release {
             minifyEnabled false
@@ -33,20 +37,20 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:26.1.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    implementation 'com.android.support:design:26.1.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
+    implementation 'com.google.android.material:material:1.0.0'
     testImplementation 'junit:junit:4.12'
-    implementation 'com.android.support:support-v4:26.1.0'
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
 
 
-    implementation 'com.android.support:cardview-v7:21.0.+'
+    implementation 'androidx.cardview:cardview:1.0.0'
 
     //external library
     //arcgis
-    implementation 'com.esri.arcgisruntime:arcgis-android:100.2.1'
+    implementation 'com.esri.arcgisruntime:arcgis-android:100.13.0'
     //bottom navigation view
     implementation 'com.aurelhubert:ahbottomnavigation:2.1.0'
 
@@ -55,13 +59,13 @@ dependencies {
     implementation 'com.google.firebase:firebase-database:16.0.1'
     implementation 'com.google.firebase:firebase-auth:16.0.1'
 
-    implementation 'com.google.android.gms:play-services-auth:15.0.1'
+    implementation 'com.google.android.gms:play-services-auth:20.1.0'
 
     implementation 'com.github.PhilJay:MPAndroidChart:v3.0.3'
 
 
-
 }
+
 
 
 
